+++ app/src/full/java/com/driot/bookplayer/podcasts/PodcastHelper.java
@@ -575,7 +575,7 @@ public class PodcastHelper {
 
     public static void onPodcastClick(Context context, DisplayableEpisode ep, Podcast podcast, String caller) {
         String cover = ep.image == null || ep.image.isEmpty() ? podcast.image : ep.image;
-        int trackId = (ep.id != null) ? (int) (long) ep.id : -1;
+        long trackId = (ep.id != null) ? ep.id : -1;
         StartPlayHelper.playStream(context, Var.PLAY_MODE_PODCAST, ep.enclosureUrl, trackId, ep.title,
                 cover, caller);
     }
@@ -646,7 +646,7 @@ public class PodcastHelper {
                 imagePath);
     }
 
-    public static void addSecondToTimeListened(Context context, int trackId) {
+    public static void addSecondToTimeListened(Context context, long trackId) {
         AppDatabase db = AppDatabase.getDatabase(context.getApplicationContext());
         db.episodeDao().addSecondToTimeListened(trackId);
     }
